requirements.txt
flask==0.10.1
flask-sqlalchemy==2.0
flask-migrate==1.3.1
flask-login==0.2.11
celery[redis]==3.1.17
pyyaml==3.11
requests[security]==2.7.0
raven[flask]==5.4.4
